renamed navBar to NavBarNEW
