+++ libs/spark/src/NavBarNEW.tsx
